standardized LinkPreview implementation, finalized changes
