+++ app/dub.co/tools/metatags/content.tsx
@@ -1,24 +1,19 @@
 "use client";
 
-import { useEffect, useMemo, useRef, useState } from "react";
+import { useState } from "react";
 import { useDebounce } from "use-debounce";
 import useSWR from "swr";
-import { Copy, Photo, Tick } from "@/components/shared/icons";
-import { LoadingCircle } from "#/ui/icons";
-import { fetcher, getDomainWithoutWWW, getUrlFromString } from "#/lib/utils";
+import { Copy, Tick } from "@/components/shared/icons";
+import { fetcher, getUrlFromString } from "#/lib/utils";
 import { toast } from "sonner";
-import { useRouter, useSearchParams } from "next/navigation";
+import { useSearchParams } from "next/navigation";
 
 export default function MetatagsContent() {
-  const router = useRouter();
   const searchParams = useSearchParams();
   const url = searchParams?.get("url") || "https://github.com/steven-tey/dub";
   const [debouncedUrl] = useDebounce(getUrlFromString(url), 500);
-  const hostname = useMemo(() => {
-    return getDomainWithoutWWW(debouncedUrl || "");
-  }, [debouncedUrl]);
 
-  const { data, isValidating } = useSWR<{
+  const { data } = useSWR<{
     title: string | null;
     description: string | null;
     image: string | null;
@@ -30,78 +25,8 @@ export default function MetatagsContent() {
 
   const [copied, setCopied] = useState(false);
 
-  const inputRef = useRef<HTMLInputElement>(null);
-  useEffect(() => {
-    inputRef.current?.select();
-  }, []);
-
   return (
     <>
-      <div className="w-full rounded-md shadow-sm">
-        <input
-          ref={inputRef}
-          name="url"
-          id="url"
-          type="url"
-          autoFocus
-          className="block w-full rounded-md border-gray-300 text-sm text-gray-900 placeholder-gray-300 focus:border-gray-500 focus:outline-none focus:ring-gray-500"
-          placeholder="Enter your URL"
-          defaultValue={url}
-          onChange={(e) =>
-            router.replace(
-              `/tools/metatags${
-                e.target.value.length > 0 ? `?url=${e.target.value}` : ""
-              }`,
-            )
-          }
-          aria-invalid="true"
-        />
-      </div>
-
-      <div className="relative overflow-hidden rounded-md border border-gray-300 bg-gray-50">
-        {isValidating && (
-          <div className="absolute flex h-[250px] w-full flex-col items-center justify-center space-y-4 border-b border-gray-300 bg-gray-50">
-            <LoadingCircle />
-          </div>
-        )}
-        {image ? (
-          <img
-            src={image}
-            alt="Preview"
-            className="h-[250px] w-full border-b border-gray-300 object-cover"
-          />
-        ) : (
-          <div className="flex h-[250px] w-full flex-col items-center justify-center space-y-4 border-b border-gray-300">
-            <Photo className="h-8 w-8 text-gray-400" />
-            <p className="text-sm text-gray-400">
-              Enter a link to generate a preview.
-            </p>
-          </div>
-        )}
-        <div className="grid gap-1 bg-white p-3 text-left">
-          {hostname ? (
-            <p className="text-sm text-[#536471]">{hostname}</p>
-          ) : (
-            <div className="mb-1 h-4 w-24 rounded-md bg-gray-100" />
-          )}
-          {title ? (
-            <h3 className="truncate text-sm font-medium text-[#0f1419]">
-              {title}
-            </h3>
-          ) : (
-            <div className="mb-1 h-4 w-full rounded-md bg-gray-100" />
-          )}
-          {description ? (
-            <p className="line-clamp-2 text-sm text-[#536471]">{description}</p>
-          ) : (
-            <div className="grid gap-2">
-              <div className="h-4 w-full rounded-md bg-gray-100" />
-              <div className="h-4 w-48 rounded-md bg-gray-100" />
-            </div>
-          )}
-        </div>
-      </div>
-
       <button
         className="hover:bg/black-[0.08] group relative flex cursor-copy items-center space-x-5 rounded-full bg-black/5 py-2.5 pl-5 pr-3 transition-all"
         onClick={() => {
